refactor: 💡 dump components
