+++ src/app/dumper/@contents/page.tsx
@@ -1,9 +1,9 @@
 import { LoadingStack } from "@/components/stack/loading-stack";
 import { Separator } from "@/components/ui/separator";
 import { checkPostPermission } from "@/features/auth/utils/role";
+import { AddContentsProvider } from "@/features/dump/components/add-contents-provider";
 import { AddFormLoading } from "@/features/dump/components/add-form-loading";
-import { ContentsAddProvider } from "@/features/dump/components/contents-add-provider";
-import { ContentsContents } from "@/features/dump/components/contents-contents";
+import { ContentsStackProvider } from "@/features/dump/components/contents-stack-provider";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -14,12 +14,12 @@ export default async function Page() {
 		<>
 			{hasPostPermission && (
 				<Suspense fallback={<AddFormLoading showCategory={false} />}>
-					<ContentsAddProvider />
+					<AddContentsProvider />
 				</Suspense>
 			)}
 			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 			<Suspense fallback={<LoadingStack />}>
-				<ContentsContents />
+				<ContentsStackProvider />
 			</Suspense>
 		</>
 	);
