+++ src/app/dumper/@news/page.tsx
@@ -2,8 +2,8 @@ import { LoadingStack } from "@/components/stack/loading-stack";
 import { Separator } from "@/components/ui/separator";
 import { checkPostPermission } from "@/features/auth/utils/role";
 import { AddFormLoading } from "@/features/dump/components/add-form-loading";
-import { NewsAddProvider } from "@/features/dump/components/news-add-provider";
-import { NewsContents } from "@/features/dump/components/news-contents";
+import { AddNewsProvider } from "@/features/dump/components/add-news-provider";
+import { NewsStackProvider } from "@/features/dump/components/news-stack-provider";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -15,12 +15,12 @@ export default async function Page() {
 		<>
 			{hasPostPermission && (
 				<Suspense fallback={<AddFormLoading showCategory />}>
-					<NewsAddProvider />
+					<AddNewsProvider />
 				</Suspense>
 			)}
 			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 			<Suspense fallback={<LoadingStack />}>
-				<NewsContents />
+				<NewsStackProvider />
 			</Suspense>
 		</>
 	);
