+++ src/features/dump/components/add-contents-form.tsx
@@ -3,60 +3,43 @@ import { SubmitButton } from "@/components/submit-button";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
-import { addNews } from "@/features/dump/actions/add-news";
-import { newsContext } from "@/features/dump/stores/news-context";
+import { addContents } from "@/features/dump/actions/add-contents";
+import { contentsContext } from "@/features/dump/stores/contents-context";
 import { useToast } from "@/hooks/use-toast";
-import type { Categories } from "@prisma/client";
 import { ClipboardPasteIcon } from "lucide-react";
-import { useRef, useState } from "react";
+import { useRef } from "react";
 import { useSetRecoilState } from "recoil";
 
-type Props = {
-	categories: Omit<Categories, "createdAt" | "updatedAt" | "userId">[];
-};
-
-const NEW_CATEGORY_VALUE = "new"; // 新規の場合のcategoryのvalue
-
-export function NewsAddForm({ categories }: Props) {
+export function AddContentsForm() {
 	const titleInputRef = useRef<HTMLInputElement>(null);
 	const quoteInputRef = useRef<HTMLTextAreaElement>(null);
 	const urlInputRef = useRef<HTMLInputElement>(null);
 
 	const { toast } = useToast();
 
-	const setQueuedContents = useSetRecoilState(newsContext);
-
-	const [newCategoryInputOpen, setNewCategoryInputOpen] = useState(false);
+	const setQueuedContents = useSetRecoilState(contentsContext);
 
 	const formAction = async (formData: FormData) => {
-		const state = await addNews(formData);
-		if (!state.success) {
+		const response = await addContents(formData);
+		if (!response.success) {
 			toast({
 				variant: "destructive",
-				description: state.message,
+				description: response.message,
 			});
 			return;
 		}
-		setQueuedContents((previousData) => [state.data, ...(previousData ?? [])]);
+		setQueuedContents((previousData) => [
+			response.data,
+			...(previousData ?? []),
+		]);
 		toast({
 			variant: "default",
-			description: state.message,
+			description: response.message,
 		});
 		if (titleInputRef.current) titleInputRef.current.value = "";
 		if (quoteInputRef.current) quoteInputRef.current.value = "";
 		if (urlInputRef.current) urlInputRef.current.value = "";
-	};
-
-	const handleSelectValueChange = (value: string) => {
-		setNewCategoryInputOpen(value === NEW_CATEGORY_VALUE);
 	};
 
 	const handlePasteClick = async () => {
@@ -67,38 +50,6 @@ export function NewsAddForm({ categories }: Props) {
 	return (
 		// MEMO: experimental feature of using form actions
 		<form action={formAction} className="space-y-4 p-4">
-			<div className="space-y-1">
-				<Label htmlFor="category">カテゴリー</Label>
-				<Select
-					name="category"
-					required
-					onValueChange={handleSelectValueChange}
-				>
-					<SelectTrigger>
-						<SelectValue />
-					</SelectTrigger>
-					<SelectContent>
-						{categories.map((category) => (
-							<SelectItem value={String(category.id)} key={category.id}>
-								{category.name}
-							</SelectItem>
-						))}
-						<SelectItem value={NEW_CATEGORY_VALUE} key={NEW_CATEGORY_VALUE}>
-							その他
-						</SelectItem>
-					</SelectContent>
-				</Select>
-			</div>
-			{newCategoryInputOpen && (
-				<div className="space-y-1">
-					<Input
-						id="new_category"
-						name="new_category"
-						required
-						autoComplete="off"
-					/>
-				</div>
-			)}
 			<div className="space-y-1">
 				<Label htmlFor="title">タイトル</Label>
 				<Input
@@ -126,8 +77,8 @@ export function NewsAddForm({ categories }: Props) {
 						name="url"
 						type="url"
 						inputMode="url"
-						ref={urlInputRef}
 						autoComplete="off"
+						ref={urlInputRef}
 						required
 					/>
 					<Button variant="ghost" type="button" onClick={handlePasteClick}>
