+++ src/features/dump/components/add-contents-provider.tsx
@@ -1,0 +1,10 @@
+import { AddContentsForm } from "./add-contents-form";
+
+export function AddContentsProvider() {
+	try {
+		return <AddContentsForm />;
+	} catch (error) {
+		console.error("Unexpected error.", error);
+		return <></>;
+	}
+}
