+++ src/features/dump/components/add-news-form.tsx
@@ -3,40 +3,60 @@ import { SubmitButton } from "@/components/submit-button";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import {
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+} from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
-import { addContents } from "@/features/dump/actions/add-contents";
-import { contentsContext } from "@/features/dump/stores/contents-context";
+import { addNews } from "@/features/dump/actions/add-news";
+import { newsContext } from "@/features/dump/stores/news-context";
 import { useToast } from "@/hooks/use-toast";
+import type { Categories } from "@prisma/client";
 import { ClipboardPasteIcon } from "lucide-react";
-import { useRef } from "react";
+import { useRef, useState } from "react";
 import { useSetRecoilState } from "recoil";
 
-export function ContentsAddForm() {
+type Props = {
+	categories: Omit<Categories, "createdAt" | "updatedAt" | "userId">[];
+};
+
+const NEW_CATEGORY_VALUE = "new"; // 新規の場合のcategoryのvalue
+
+export function AddNewsForm({ categories }: Props) {
 	const titleInputRef = useRef<HTMLInputElement>(null);
 	const quoteInputRef = useRef<HTMLTextAreaElement>(null);
 	const urlInputRef = useRef<HTMLInputElement>(null);
 
 	const { toast } = useToast();
 
-	const setQueuedContents = useSetRecoilState(contentsContext);
+	const setQueuedContents = useSetRecoilState(newsContext);
+
+	const [newCategoryInputOpen, setNewCategoryInputOpen] = useState(false);
 
 	const formAction = async (formData: FormData) => {
-		const state = await addContents(formData);
-		if (!state.success) {
+		const response = await addNews(formData);
+		if (!response.success) {
 			toast({
 				variant: "destructive",
-				description: state.message,
+				description: response.message,
 			});
 			return;
 		}
-		setQueuedContents((previousData) => [state.data, ...(previousData ?? [])]);
+		setQueuedContents((previousData) => [response.data, ...(previousData ?? [])]);
 		toast({
 			variant: "default",
-			description: state.message,
+			description: response.message,
 		});
 		if (titleInputRef.current) titleInputRef.current.value = "";
 		if (quoteInputRef.current) quoteInputRef.current.value = "";
 		if (urlInputRef.current) urlInputRef.current.value = "";
+	};
+
+	const handleSelectValueChange = (value: string) => {
+		setNewCategoryInputOpen(value === NEW_CATEGORY_VALUE);
 	};
 
 	const handlePasteClick = async () => {
@@ -47,6 +67,38 @@ export function ContentsAddForm() {
 	return (
 		// MEMO: experimental feature of using form actions
 		<form action={formAction} className="space-y-4 p-4">
+			<div className="space-y-1">
+				<Label htmlFor="category">カテゴリー</Label>
+				<Select
+					name="category"
+					required
+					onValueChange={handleSelectValueChange}
+				>
+					<SelectTrigger>
+						<SelectValue />
+					</SelectTrigger>
+					<SelectContent>
+						{categories.map((category) => (
+							<SelectItem value={String(category.id)} key={category.id}>
+								{category.name}
+							</SelectItem>
+						))}
+						<SelectItem value={NEW_CATEGORY_VALUE} key={NEW_CATEGORY_VALUE}>
+							その他
+						</SelectItem>
+					</SelectContent>
+				</Select>
+			</div>
+			{newCategoryInputOpen && (
+				<div className="space-y-1">
+					<Input
+						id="new_category"
+						name="new_category"
+						required
+						autoComplete="off"
+					/>
+				</div>
+			)}
 			<div className="space-y-1">
 				<Label htmlFor="title">タイトル</Label>
 				<Input
@@ -74,8 +126,8 @@ export function ContentsAddForm() {
 						name="url"
 						type="url"
 						inputMode="url"
-						autoComplete="off"
 						ref={urlInputRef}
+						autoComplete="off"
 						required
 					/>
 					<Button variant="ghost" type="button" onClick={handlePasteClick}>
