--- src/features/dump/components/contents-add-provider.tsx
@@ -1,10 +1,0 @@
-import { ContentsAddForm } from "./contents-add-form";
-
-export function ContentsAddProvider() {
-	try {
-		return <ContentsAddForm />;
-	} catch (error) {
-		console.error("Unexpected error.", error);
-		return <></>;
-	}
-}
