+++ src/features/dump/components/contents-stack-provider.tsx
@@ -1,20 +1,19 @@
-import { getSelfUnexportedNews } from "@/apis/prisma/fetch-news";
+import { getSelfUnexportedContents } from "@/apis/prisma/fetch-contents";
 import { StatusCodeView } from "@/components/status-code-view";
-import { NewsStack } from "./news-stack";
+import { ContentsStack } from "./contents-stack";
 
-export async function NewsContents() {
+export async function ContentsStackProvider() {
 	try {
-		const unexportedNews = await getSelfUnexportedNews();
+		const unexportedContents = await getSelfUnexportedContents();
 
 		return (
-			<NewsStack
-				news={unexportedNews.map((d) => {
+			<ContentsStack
+				contents={unexportedContents.map((d) => {
 					return {
 						id: d.id,
 						title: d.title,
 						quote: d.quote,
 						url: d.url,
-						category: d.Category.name,
 					};
 				})}
 			/>
